+++ src/main/java/ddg/walking_rabbit/global/domain/entity/ChatRecordEntity.java
@@ -1,4 +1,4 @@
-package ddg.walking_rabbit.message.entity;
+package ddg.walking_rabbit.global.domain.entity;
 
 import ddg.walking_rabbit.user.entity.UserEntity;
 import jakarta.persistence.*;
