+++ src/main/java/ddg/walking_rabbit/global/domain/entity/ConversationEntity.java
@@ -1,10 +1,8 @@
-package ddg.walking_rabbit.message.entity;
+package ddg.walking_rabbit.global.domain.entity;
 
 import ddg.walking_rabbit.user.entity.UserEntity;
 import jakarta.persistence.*;
 import lombok.*;
-
-import java.time.LocalDateTime;
 
 @Entity
 @NoArgsConstructor
