+++ src/main/java/ddg/walking_rabbit/global/domain/entity/MessageEntity.java
@@ -1,0 +1,30 @@
+package ddg.walking_rabbit.global.domain.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+public class MessageEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long messageId;
+
+    @ManyToOne
+    @JoinColumn(nullable = false)
+    private ConversationEntity conversation;
+
+    @Enumerated(EnumType.STRING)
+    private Role role;
+
+    @Enumerated(EnumType.STRING)
+    private ContentType contentType;
+
+    private String content;
+
+}
