+++ src/main/java/ddg/walking_rabbit/global/domain/entity/Role.java
@@ -1,0 +1,5 @@
+package ddg.walking_rabbit.global.domain.entity;
+
+public enum Role {
+    USER, ASSISTANT
+}
