+++ src/main/java/ddg/walking_rabbit/global/domain/repository/ConversationRepository.java
@@ -1,6 +1,6 @@
-package ddg.walking_rabbit.message.repository;
+package ddg.walking_rabbit.global.domain.repository;
 
-import ddg.walking_rabbit.message.entity.ConversationEntity;
+import ddg.walking_rabbit.global.domain.entity.ConversationEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ConversationRepository extends JpaRepository<ConversationEntity, Long> {
