+++ src/main/java/ddg/walking_rabbit/global/domain/repository/MissionRepository.java
@@ -1,11 +1,8 @@
-package ddg.walking_rabbit.message.repository;
+package ddg.walking_rabbit.global.domain.repository;
 
-import ddg.walking_rabbit.message.dto.ChatStartDto;
-import ddg.walking_rabbit.message.entity.MissionEntity;
+import ddg.walking_rabbit.global.domain.entity.MissionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 @Repository
 public interface MissionRepository extends JpaRepository<MissionEntity, Long> {
