--- src/main/java/ddg/walking_rabbit/message/entity/ContentType.java
@@ -1,5 +1,0 @@
-package ddg.walking_rabbit.message.entity;
-
-public enum ContentType {
-    PHOTO, TEXT
-}
