--- src/main/java/ddg/walking_rabbit/message/entity/MissionEntity.java
@@ -1,29 +1,0 @@
-package ddg.walking_rabbit.message.entity;
-
-import ddg.walking_rabbit.user.entity.UserEntity;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDate;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-public class MissionEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long missionId;
-
-    @ManyToOne
-    @JoinColumn(nullable = false)
-    private UserEntity user;
-
-    private String content;
-
-    private LocalDate missionDate;
-
-}
