--- src/main/java/ddg/walking_rabbit/message/entity/Role.java
@@ -1,5 +1,0 @@
-package ddg.walking_rabbit.message.entity;
-
-public enum Role {
-    USER, ASSISTANT
-}
