+++ src/main/java/ddg/walking_rabbit/message/service/MessageService.java
@@ -1,16 +1,15 @@
 package ddg.walking_rabbit.message.service;
 
 import com.google.cloud.storage.*;
+import ddg.walking_rabbit.global.domain.entity.*;
 import ddg.walking_rabbit.message.dto.ChatResponseDto;
 import ddg.walking_rabbit.message.dto.ChatStartDto;
-import ddg.walking_rabbit.message.entity.*;
-import ddg.walking_rabbit.message.repository.ConversationRepository;
-import ddg.walking_rabbit.message.repository.MessageRepository;
-import ddg.walking_rabbit.message.repository.MissionRepository;
+import ddg.walking_rabbit.global.domain.repository.ConversationRepository;
+import ddg.walking_rabbit.global.domain.repository.MessageRepository;
+import ddg.walking_rabbit.global.domain.repository.MissionRepository;
 import ddg.walking_rabbit.user.entity.UserEntity;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.tomcat.util.http.fileupload.FileUploadException;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,7 +17,6 @@ import org.springframework.web.client.RestTemplate;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.io.InputStream;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 import java.util.UUID;
